requirements.txt
Numpy
tqdm
tensorflow
pillow
pydot
matplotlib
